Move the contents of detail/tuple.hpp into detail/tuple_utility.hpp Eliminate detail/tuple.hpp
